+++ test/HylianShield/Tests/Validator/CoreArray/NumericTest.php
@@ -4,7 +4,6 @@
  *
  * @package HylianShield
  * @subpackage Test
- * @copyright 2014 Noë Snaterse.
  */
 namespace HylianShield\Tests\Validator\CoreArray;
 
@@ -30,7 +29,6 @@ class NumericTest extends \HylianShield\Tests\Validator\TestBase
         array('0123456789', false),
         array('', false),
         array('€αβγδε', false),
-        array(0123456789, false),
         array(0.123456789, false),
         array(null, false),
         array(0, false),