+++ test/HylianShield/Tests/Validator/Integer/NegativeTest.php
@@ -4,12 +4,9 @@
  *
  * @package HylianShield
  * @subpackage Test
- * @copyright 2014 Jan-Marten "Joh Man X" de Boer
  */
 
 namespace HylianShield\Tests\Validator\Integer;
-
-use \HylianShield\Validator\Integer\Negative;
 
 /**
  * Negative test.