+++ test/HylianShield/Tests/Validator/IntegerTest.php
@@ -4,7 +4,6 @@
  *
  * @package HylianShield
  * @subpackage Test
- * @copyright 2014 Jan-Marten "Joh Man X" de Boer
  */
 
 namespace HylianShield\Tests\Validator;
@@ -33,7 +32,6 @@ class IntegerTest extends \HylianShield\Tests\Validator\TestBase
         array('0123456789', false),
         array('', false),
         array('€αβγδε', false),
-        array(0123456789, true),
         array(0.123456789, false),
         array(null, false),
         array(0, true)
@@ -46,7 +44,9 @@ class IntegerTest extends \HylianShield\Tests\Validator\TestBase
     public function testDefaultNotZero()
     {
         $validator = $this->validatorClass;
+        /** @var \HylianShield\Validator\Integer $validator */
         $validator = new $validator(0, 0);
+
         $this->assertTrue($validator->validate(0));
         $this->assertFalse($validator->validate(1));
         $this->assertFalse($validator->validate(-1));